enable pixel snapping - removes world seams
