+++ C7/C7Game.tscn
@@ -68,6 +68,10 @@ script = ExtResource("1")
 anchors_preset = 15
 anchor_right = 1.0
 anchor_bottom = 1.0
+offset_left = -574.0
+offset_top = -322.0
+offset_right = -574.0
+offset_bottom = -322.0
 grow_horizontal = 2
 grow_vertical = 2
 theme = ExtResource("10")
