+++ C7/PlayerCamera.cs
@@ -5,10 +5,21 @@ public partial class PlayerCamera : Camera2D
 {
 	private readonly float maxZoom = 2.0f;
 	private readonly float minZoom = 0.2f;
-	private float zoomFactor = 1.0f;
+	public float zoomFactor {get; private set; } = 1.0f;
 
 	public override void _Ready() {
 		base._Ready();
+		scaleZoom(zoomFactor);
+	}
+
+	public void scaleZoom(float factor) {
+		zoomFactor = zoomFactor * factor;
+		zoomFactor = Mathf.Clamp(zoomFactor, minZoom, maxZoom);
+		Zoom = Vector2.One * zoomFactor;
+	}
+
+	public void setZoom(float factor) {
+		zoomFactor = Mathf.Clamp(factor, minZoom, maxZoom);
 		Zoom = Vector2.One * zoomFactor;
 	}
 
@@ -17,8 +28,7 @@ public partial class PlayerCamera : Camera2D
 			Position -= mm.Relative / Zoom;
 		}
 		if (@event is InputEventMagnifyGesture mg) {
-			zoomFactor = Mathf.Clamp(zoomFactor * mg.Factor, minZoom, maxZoom);
-			Zoom = Vector2.One * zoomFactor;
+			scaleZoom(mg.Factor);
 		}
 	}
