+++ C7/tests/TestUnit.cs
@@ -1,12 +1,8 @@
 using Godot;
-using System;
-using ConvertCiv3Media;
 using C7GameData;
 
 public partial class TestUnit : Node2D
 {
-
-	// Called when the node enters the scene tree for the first time.
 	public override void _Ready()
 	{
 		AnimationManager manager = new AnimationManager(null);
@@ -30,10 +26,5 @@ public partial class TestUnit : Node2D
 		cursor.Position = new Vector2(120, 30);
 		cursor.Play("cursor");
 		AddChild(cursor);
-	}
-
-	// Called every frame. 'delta' is the elapsed time since the previous frame.
-	public override void _Process(double delta)
-	{
 	}
 }
